requirements.txt
requests
wasabi
python-dotenv
goldenverba